+++ modules/sys/partials.js
@@ -42,6 +42,7 @@ export default function preloadHandlebarsTemplates() {
 
         'systems/hackmaster5e/templates/chat/top.hbs',
         'systems/hackmaster5e/templates/chat/trauma.hbs',
+        'systems/hackmaster5e/templates/chat/skill.hbs',
         'systems/hackmaster5e/templates/chat/spell.hbs',
 
         'systems/hackmaster5e/templates/item/parts/description.hbs',