+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Can specify skill check difficulty.
+- Language hotbar macros.
 ### Changed
 - Attack dialogs now specify weapon speed and only offer range mods for ranged weapons.